#1 fix: integer ids in output after SDK retype
The SDK now types ids as i64, so JSON renders them as integers
(1873345 instead of 1873345.0). Adapt twc-rs: fmt_id is generic over
Display (handles both i64 ids and f64 cpu/ram), the balancer preset_id and
rule-id comparison use integers, and the TUI account id is cast to f64 for
its widget. fmt_id tests updated to integer formatting.

+++ src/main.rs
@@ -1198,7 +1198,7 @@ async fn refresh_all(
     let mut balance = String::new();
 
     if let Ok(resp) = account_res {
-        account_id = resp.status.company_info.id;
+        account_id = resp.status.company_info.id as f64;
         login = resp.status.login.clone().unwrap_or_default();
     } else {
         has_error = true;
